rename site header and footer components
