+++ src/components/site-footer.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import ThemeToggle from "./theme-toggle";
 
-export default function Footer() {
+export default function SiteFooter() {
   const currentYear = new Date().getFullYear();
   return (
     <footer className="w-full h-48 md:h-36 mx-auto my-0 px-8 md:px-4 flex flex-col md:flex-row justify-evenly md:justify-between items-start md:items-center font-mono text-xs">
