+++ tailwind.config.ts
@@ -1,7 +1,6 @@
 import type { Config } from "tailwindcss";
 import { fontFamily } from "tailwindcss/defaultTheme";
 
-// /** @type {import('tailwindcss').Config} */
 export default {
   darkMode: ["class"],
   content: [
